+++ src/gnu/CMakeLists.txt
@@ -1,22 +1,15 @@
 add_library(
   nrngnu STATIC
-  DiscUnif.cpp
   Erlang.cpp
-  Geom.cpp
   HypGeom.cpp
-  LogNorm.cpp
   MCellRan4RNG.cpp
   mcran4.cpp
-  NegExp.cpp
-  Normal.cpp
   NrnRandom123RNG.cpp
   nrnran123.cpp
-  Poisson.cpp
   Rand.cpp
   Random.cpp
   RndInt.cpp
   RNG.cpp
-  Uniform.cpp
   Weibull.cpp)
 set_property(TARGET nrngnu PROPERTY POSITION_INDEPENDENT_CODE ON)
 target_include_directories(nrngnu SYSTEM PRIVATE "${PROJECT_SOURCE_DIR}/external/Random123/include")
