+++ src/ivoc/ivocvect.cpp
@@ -97,7 +97,7 @@ static double dmaxint_ = 9007199254740992;
 
 // definition of random numer generator
 #include "Rand.hpp"
-#include <Uniform.h>
+#include "distributions.hpp"
 
 #if HAVE_IV
 #include "utility.h"
